First draft specification of the SMRRs
The main limitation of this specification is the SMRRs are hardcoded
to designate the SMRAM and only the SMRAM, whether the true
implementation of Intel is more flexible. This is not a big issue here
as we only want to have a quick example. It will be possible to have a
more flexible way to deal with it latter.