feature: Add the [Use] typeclass to generalize extensible effects
The [row] type and the [HasEffects] typeclass introduced recently
allows for implementing extensible effects using FreeSpec. The truth
is, the [<+>] composition operator can also be use to that extend, and
we can easily imagine others.

This is why this commit introduce the [Use] typeclass to postpone the
choice of interface aggregator.