feature: Change the equality predicate for Program
The last version of program_eq had a major issue: it was not possible
to easily exploit hypotheses of the form `program_eq p q`. For
instance, I was not able to prove that for two equal/equivalent
programs, the resulting abstract state was the same. This new version
of `program_eq` focuses on program rewriting.