feature: Provide a novel way to execute [Program] instance
This patch features a new plugin for the Coq theorem prover which
notably introduces the `Exec' keyword. The goal of `Exec' is to act as
an handler for certain supported interfaces (currently, only
`Console.i'). This effectively turns `coqc' into a Gallina interpreter
with side-effects.

Although the current implementation is far from feature-completed, it
already works pretty. The interested reader can have a look at the
`exec/examples' directory.