stdlib: Add [CommandLine] interface to get command line args
In order to get command line arguments from the Coq side, the program
must be run with the `fsrun` script (in the `run` directory). It uses
environment variables to transmit the args given by the user. It is also
no longer necessary to add `Exec xxx.` at the end of the program file,
since `fsrun` will append `Exec main.` to the file when calling `coqc`.
`fsrun` will also remove the first line of the program file if it begins
with `#!`, so it is possible to add `#!/usr/bin/env fsrun` and run the
program with `./Prog.v`, if `fsrun` is accesible in the environment path.