refactor: Deal with opaque program in [prove_program] tactic
By introducing a new predicate [run] to reason about execution of
correct programs by compliant handler, it becomes possible to reason
about opaque program. The [prove_program] now relies on a dedicated
lemmas to abstract away both the result of the program and the new
witness state.

This allows for making programs opaque, in order to avoid having to
prove the same properties multiple times.