feature: Add a convenient notation to reduce [Use] verbosity
This patch introduces the notation [ix :| i1, i2, i3] (“ix provides
i1, i2 and i3”) to be used in place of [Use i1 ix, Use i2 ix, Use i3
ix].

Thus, where we had to write

    Definition my_program {ix} `{Use Console.i ix, Use Debug.i ix}
      : Program ix A :=
      (* ... *)

We can now write

    Definition my_program {ix} `{ix :| Console.i, Debug.i}
      : Program ix A :=
      (* ... *)

As a result, Program types are (at least to the author’s opinion)
easier to read.  The main inspiration for this notation is the
PureScipt so-called “row of effects”.  One major feature of this
notation is that the order of the row is not important.