fix: Change the definition of [program_equiv]
We need this change to be able to prove that:

    p == q -> trustworthy_program c w p -> trustworthy_program c w q

This way, when we have to prove

    trustworthy_program c w (bind (bind p f) g)

We can rewrite this equation using the bind associativity law.

Note that, although I have never been able to prove it, I do believe
the two definitions (the old and the new one) are basically
equivalent.