refactor: Rename [program] into [impure], and document the library
This may probably be a very controversial change. This commit rename
the [program] monad into [impure], talks about impure computations
rather than programs with effects and impure primitives rather than
effects.

One reason behind this choice is to avoid any confusion with the
Program framework of Coq. The main reason, however, is that our monad
is more limited compare to algebraic effects, as they are described in
the literature.

Besides, we provide more documentation about the FreeSpec.Core.Impure
library.