refactor: Use [Coqlib.ref_lib] to locate Coq constants
In coq-8.10, most of the [Coqlib] module has been deprecated in favor
of [Coqlib.ref_lib].  The way Coq constants definitions are requested
in Ocaml changes, and now requires these constants to be registered.

We update the freespec-exec plugin to use this new approach.

An important consequence of this change in Coq is for FreeSpec users:
they will have to register their interfaces too.