fix: Add the missing [Proper] instance for [trustworthy_run]
Without this instance, the [impure_bind_assoc] lemma cannot be used
when reasoning about trustworthy run.  The [unroll_impure_run] tactic
actually is trying to do just that when it finds an impure computation
of the form [impure_bind (impure_bind _ _) _], and therefore was
failing with a lovely setoid-related error message.