feature: Add a first prototype [FILES] interface
This patch introduces many useful utility features which were
necessary to implement a robust [FILES] interface.

From [Exec] perspective, it adds several functions to manipulate
[prod] and [sum]. They have the particularity to be polymorphic, where
previous types handled by [Exec] where not. The interface could be a
bit easier to use, but the current implementation is straightforward
enough.

From the [Stdlib] perspective, we add the [RAISE] interface, to easily
specify “computation shortcuts.” The main idea is that an impure
computation of type [impure (RAISE a) b] is a computation which is
expected to produce a term of type [b], but may also be interrupted
while doing it, and in such a case produces a value of type [b] to be
handled by the callee.

We use this mechanics to deal with primitives of the [FILES]
interface. The resulting interface is pretty straightforward to
use. See for instance the following code snippet:

    recover
        do var fd <- try_open "index.html" in
           var s <- try_file_size fd in

           var txt <- try_read fd 10 in
           echo (snd txt);

           close fd;
           echo "it worked"
         end
       with err of files_err =>
         echo "it did not worked."
    end