chore: Use the latest version of the dune configuration syntax
Dune configuration file syntax is versioned, and with this commit we
explicitly use the 2.4 version of dune. We do that because the Coq
support of dune remains experimental, and therefore is subject to
change. For instance, the public_name stanza has been deprecated, and
there is no reason to continue using it. Besides, dune will soon add
support for compositional build of Coq libraries and we really want to
use that.