feature: Add a [HEAP] interface and the dedicated Exec handlers
This interface allows for storing Coq terms inside a so-called heap.
Besides, it extends the Exec vernacular command with dedicated
handlers. These handlers use a shared, mutable hash table to store
the terms, and a stream of integers to generate fresh references.

Wrt. the Exec command, note that a reference [k] used at a given time
for accessing a term [t] can later be used for accessing a term [t']
of totally different types if [k] has been freed (by means of the
[Destruct] primitive). Therefore, using a freed reference is
unsafe. This is similar to the use-after-free safety/security issue of
most C programs.

See plugins/exec/heap.mli to see how a OCaml plugin can create a
reference, and theories/Exec/Heap.v to see how to manipulate said
references from an computations defined in [impure].