feature: Provide an opaque resources handler
The main idea behind this patch is to provide a so-called “resources
manager.” The expected users are primitives handler which manipulates
values which cannot be encoded in Coq. A good example of this is the
primitives of the [FILES] interface. Said primitives manipulates file
descriptors, implemented in OCaml as an opaque type.

Rather than encoding these values directly inside Coq terms, we
inject a primitive integers to be used as a key to manipulate the
actual value, as stored inside the resources manager. Therefore, going
back to the [FILES] interface, the [Open] primitive handler will
insert the opened file descriptor inside the resources manager, then
returns the key used to store it. The [Read] and [Write] primitives
handlers retrieve the file descriptors from the key, and the [Close]
primitive handler removes the key from the resources manager.

From the Coq side, the good approach is to define the
[file_descriptor] type as an axiom. By doing that, we know a user of
the [FILES] interface will not be able to forge or alter a file
descriptor.