refactor: Use Coq terms in place of OCaml integers for indirection
coq-freespec-exec package allows to work with OCaml-like references
(Heap) and opaque values that we cannot encode in Coq (Resources).  In
a nutshell, these two modules wrap hash tables to store values, and
inject in a Coq term a primitive integers which acts as keys to
retrieve the value.

Previous implementations of these module were leaking this
implementation details. We now abstract the type of the keys by
returning a Constr.t value. This make the use of said keys more
straightforward for plugins.