refactor: Migrate to a Hoare Logic reasoning framework
We implement the third iteration of the FreeSpec’s formalism.
Following the example provided by “Dijkstra Monads for All” by
Maillard et al. (ICFP,19), we leverage FreeSpec’s contract to
implement a general morphism between [impure i] and [hoare S].

Indeed, it turns out that

    respectful_impure c p s === pre (to_hoare c p) s
    respectful_run c p s s' x === pre (to_hoare c p) s
                                    -> post (to_hoare c p) s x s'

This patch changes the formalism, but not the terminology. For
instance, the code continues to refer to “respectful impure programs.”
Updating the terminology remains an important work to be done, but
we’d rather benefit from a bit more hindsight before doing the change.

We also take the opportunity given by this important refactoring to
(1) drop coq-prelude/coq-base as a dependency and use coq-extlib
instead, (2) reorganize the module hierarchy of FreeSpec.Core, and (3)
to remove FreeSpec.Stdlib (it was never completed or usable to begin
with).