refactor: Introduce the coq-freespec-ffi package
The key idea behind this package is to provide the necessary
abstraction for FreeSpec to work with coqffi. Prior to this change, it
was split between FreeSpec.Core (the [Inject] instance) and
FreeSpec.Exec (the HEAP interface).

With this change, FreeSpec.Core only depends on ExtLib, and users are
not expected to have coqffi installed. Besides, FreeSpec.Exec and
FreeSpec.FFI becomes very lightweight packages, with one very specific
feature at their core.