refactor: Prefer a rc file to the use of the `global' attribute
Prior to this patch, we were configuring Coq once and for all in the
[FreeSpec.Core.Init] module to avoid code duplication. For instance,
the command [Generalizable All Variables] was called with the
`global' attribute. This was too opinionated, since it would mean
FreeSpec’s users would have to deal with our configuration choices.

To avoid code duplication, we now rely on a dedicated `build.v' script
and the `-init-file' command-line argument of coqtop. We believe this
is cleaner.